requirements.txt
Django==4.1.1
mysqlclient==2.1.0
sqlparse==0.4.2
gunicorn==20.0.4
uvicorn==0.18.2
djangorestframework==3.14.0
pytz==2022.2.1
celery==5.2.7
django-celery-beat==2.3.0